+++ CHANGELOG.md
@@ -2,6 +2,11 @@ CHANGE LOG
 ==========
 
 
+## V0.1.1 Alpha (Upcoming)
+
+* Expose the factory class in the ioc
+
+
 ## V0.1 Alpha (27/07/2014)
 
 * Initial testing release